+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 * Refactored config prefix in metrics
 * Removed excess status labels from metrics
 
+## v3.54.3
+* Implement `String` interface for `Driver` struct
+
 ## v3.54.2
 * Added context to some internal methods for better tracing
 * Added `trace.FunctionID` helper and `FunctionID` field to trace start info's